Let's rename old engine controller too
