+++ meson.build
@@ -181,7 +181,7 @@ common_files += [
 ]
 
 files += [
-  'src/engine.cc',
+  'src/engine_classic.cc',
   'src/engine_loop.cc',
   'src/mcts/params.cc',
   'src/mcts/search.cc',
